--- com.taken_seat.payment-service/src/main/java/com/taken_seat/payment_service/application/client/TossPaymentClient.java
@@ -1,10 +1,0 @@
-package com.taken_seat.payment_service.application.client;
-
-import com.taken_seat.payment_service.application.client.dto.TossConfirmResponse;
-import com.taken_seat.payment_service.application.client.dto.TossPaymentRequest;
-
-public interface TossPaymentClient {
-	TossConfirmResponse confirmPayment(TossPaymentRequest request);
-
-	void refund(String paymentKey, Integer cancelAmount, String cancelReason);
-}
