+++ com.taken_seat.payment-service/src/main/java/com/taken_seat/payment_service/application/service/api/PaymentService.java
@@ -3,13 +3,13 @@ package com.taken_seat.payment_service.application.service.api;
 import java.util.UUID;
 
 import com.taken_seat.common_service.dto.AuthenticatedUser;
-import com.taken_seat.payment_service.application.client.dto.TossConfirmResponse;
-import com.taken_seat.payment_service.application.client.dto.TossPaymentRequest;
 import com.taken_seat.payment_service.application.dto.controller.response.PagePaymentResponseDto;
 import com.taken_seat.payment_service.application.dto.controller.response.PaymentCheckoutResponse;
 import com.taken_seat.payment_service.application.dto.controller.response.PaymentDetailResDto;
 import com.taken_seat.payment_service.application.dto.service.PaymentDto;
 import com.taken_seat.payment_service.application.dto.service.PaymentSearchDto;
+import com.taken_seat.payment_service.application.tossclient.dto.TossConfirmResponse;
+import com.taken_seat.payment_service.application.tossclient.dto.TossPaymentRequest;
 
 public interface PaymentService {
