+++ com.taken_seat.payment-service/src/main/java/com/taken_seat/payment_service/application/service/api/PaymentServiceImpl.java
@@ -15,14 +15,14 @@ import com.taken_seat.common_service.dto.AuthenticatedUser;
 import com.taken_seat.common_service.exception.customException.PaymentException;
 import com.taken_seat.common_service.exception.customException.PaymentHistoryException;
 import com.taken_seat.common_service.exception.enums.ResponseCode;
-import com.taken_seat.payment_service.application.client.TossPaymentClient;
-import com.taken_seat.payment_service.application.client.dto.TossConfirmResponse;
-import com.taken_seat.payment_service.application.client.dto.TossPaymentRequest;
 import com.taken_seat.payment_service.application.dto.controller.response.PagePaymentResponseDto;
 import com.taken_seat.payment_service.application.dto.controller.response.PaymentCheckoutResponse;
 import com.taken_seat.payment_service.application.dto.controller.response.PaymentDetailResDto;
 import com.taken_seat.payment_service.application.dto.service.PaymentDto;
 import com.taken_seat.payment_service.application.dto.service.PaymentSearchDto;
+import com.taken_seat.payment_service.application.tossclient.TossPaymentClient;
+import com.taken_seat.payment_service.application.tossclient.dto.TossConfirmResponse;
+import com.taken_seat.payment_service.application.tossclient.dto.TossPaymentRequest;
 import com.taken_seat.payment_service.domain.model.Payment;
 import com.taken_seat.payment_service.domain.model.PaymentHistory;
 import com.taken_seat.payment_service.domain.repository.PaymentHistoryRepository;
