+++ com.taken_seat.payment-service/src/main/java/com/taken_seat/payment_service/application/tossclient/TossPaymentClient.java
@@ -1,0 +1,10 @@
+package com.taken_seat.payment_service.application.tossclient;
+
+import com.taken_seat.payment_service.application.tossclient.dto.TossConfirmResponse;
+import com.taken_seat.payment_service.application.tossclient.dto.TossPaymentRequest;
+
+public interface TossPaymentClient {
+	TossConfirmResponse confirmPayment(TossPaymentRequest request);
+
+	void refund(String paymentKey, Integer cancelAmount, String cancelReason);
+}
