+++ com.taken_seat.payment-service/src/main/java/com/taken_seat/payment_service/application/tossclient/dto/TossCancelRequest.java
@@ -1,4 +1,4 @@
-package com.taken_seat.payment_service.application.client.dto;
+package com.taken_seat.payment_service.application.tossclient.dto;
 
 public record TossCancelRequest(
 	Integer cancelAmount,
