+++ com.taken_seat.payment-service/src/main/java/com/taken_seat/payment_service/presentation/controller/TossPaymentApiController.java
@@ -10,10 +10,10 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.taken_seat.payment_service.application.client.dto.TossConfirmResponse;
-import com.taken_seat.payment_service.application.client.dto.TossPaymentRequest;
 import com.taken_seat.payment_service.application.dto.controller.response.PaymentCheckoutResponse;
 import com.taken_seat.payment_service.application.service.api.PaymentService;
+import com.taken_seat.payment_service.application.tossclient.dto.TossConfirmResponse;
+import com.taken_seat.payment_service.application.tossclient.dto.TossPaymentRequest;
 
 import lombok.RequiredArgsConstructor;
